Fix syncing cache behind HTTP proxy
The default `RoundTripper` uses `http.ProxyFromEnvironment` as it uses HTTP proxies as directed by the $HTTP_PROXY and $NO_PROXY (or $http_proxy and $no_proxy) environment variables. This allows the `RoundTripper` to work behind proxies.

I assume registering custom protocol is due to internal requirements as it is not necessary when downloading files from GitHub.